+++ src/Assets/Cognigy/Utility Classes/EventArgs/FlowOutputEventArgs.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json.Linq;
+using System;
+
+namespace Cognigy
+{
+    public class FlowOutputEventArgs : EventArgs
+    {
+        public FlowOutput FlowOutput { get; private set; }
+
+        public FlowOutputEventArgs(FlowOutput flowOutput)
+        { this.FlowOutput = flowOutput; }
+    }
+}