+++ src/Assets/Cognigy/Utility Classes/Exceptions/CognigyException.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using UnityEngine;
 
 namespace Cognigy.Utility
 {
@@ -10,6 +11,7 @@ namespace Cognigy.Utility
         public CognigyException(string message)
             : base(message)
         {
+            Debug.LogError(string.Format("-- {0} {1} -- \n\n {2}","[COGNIGY.AI]", message, this.StackTrace));
         }
 
         public CognigyException(string message, Exception inner)