+++ CHANGELOG.md
@@ -38,3 +38,32 @@
 
 - AI Client
     - Added some connection options to enable multiple connections
+
+## [Version 2.0.0] (08.06.2018)
+
+### Changed
+- AI Client
+    - Updated the connection behaviour to be able to connect to COGNIGY.AI 3.0
+
+- AI Options renamed to **SocketEndpointOptions**
+    - Reduced content to the new needed fields
+
+- Renamed **RawMessage** to **CognigyInput**
+    - Now contains the necessary parameters for the COGNIGY.AI 3.0 socket connection
+    - Each message now has a *URLToken*, *userId*, *sessionId* and *source* field
+
+- Cognigy AI Window
+    - Displays now the necessary fields
+    - Removed flow settings fields since these options are configured in the socket endpoint
+    
+### Removed
+- Since the COGNIGY.AI 3.0 Update causes changes in the connection buildup, some classes are now unnecessary - the following are removed:
+    - /UtilityClasses/Connection/InitializationParameters
+    - /UtilityClasses/Connection/RequestBodyContent
+    - /UtilityClasses/Connection/ResponseBodyContent
+    - /UtilityClasses/Exceptions/CognigyRequestException
+
+- /UtilityClasses/ErrorLogger 
+- /UtilityClasses/EventArgs/StepEventArgs
+- /UtilityClasses/Response/Step
+